refactor(server): refactor server main
